refactor(dashboard): consolidate tone classification into src/lib/tone.ts
Unify 8 toneClass variants + domain-specific tone functions into a single module:
- toneClass: superset of all input aliases (ok/warn/bad)
- chainStatusTone: substring-based chain runtime classification
- sessionStatusTone: normalized session status classification
- expiryTone: ISO timestamp expiry check
- governanceToneClass: governance decisions (positive/negative/neutral)

Input keyword sets merged as union across all original implementations.
Net: +11 lines (95 removed, 106 added).
